+++ src/main/java/roomescape/controller/web/HomeController.java
@@ -1,15 +1,15 @@
-package roomescape.controller;
+package roomescape.controller.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("reservation")
-public class ReservationController {
+@RequestMapping("/")
+public class HomeController {
 
     @GetMapping
-    public String reservation() {
-        return "reservation";
+    public String home() {
+        return "home";
     }
 }
