+++ src/main/java/roomescape/controller/web/ReservationController.java
@@ -1,15 +1,15 @@
-package roomescape.controller;
+package roomescape.controller.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/")
-public class HomeController {
+@RequestMapping("reservation")
+public class ReservationController {
 
     @GetMapping
-    public String home() {
-        return "home";
+    public String reservation() {
+        return "reservation";
     }
 }
