+++ src/main/java/roomescape/global/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.global;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
+
     @ExceptionHandler({IllegalStateException.class, IllegalArgumentException.class})
     public ResponseEntity<Void> handleException(Exception e) {
         return ResponseEntity.badRequest().build();
